style context added to roc plot
